chore: refactor pages
