+++ src/app/pages/shared/registers-page/registers-page.component.html
@@ -1,0 +1,1 @@
+<p>registers-page works!</p>
